api/api_requirements.txt
# API Requirements
# api/requirements.txt

# FastAPI and server
fastapi==0.100.0
uvicorn==0.22.0
python-multipart==0.0.6
pydantic==2.0.2

# Data processing
numpy==1.23.5
pillow==9.5.0

# Machine learning
tensorflow==2.12.0
scikit-learn==1.2.2

# Utilities
python-json-logger==2.0.7
